Add Parents to Host. Introduce PluginInfo component.
Adds a ParentsFunc function to the Host component and a Parents
function to the Host core. This is used when running host detection
to determine best match when multiple plugin results are positive.

Introduces a new PluginInfo component, which is generally internal.
Components are registered when the plugin is started and the
PluginInfo component is dynamically added. It is used for reporting
what components a plugin supports so it can be registered properly