Add folder proto arg
The folder arg represents a synced folder defined in the Vagrantfile.
It maps the name of a synced folder to its configuration (which
is defined by the plugin).

For a legacy Ruby plugins the configuration may be a hash, for a
go plugin, the configuration should be defined as a proto message.