Make SSHInfo Fields optional
The successful use of ssh_info depends on it being sparsely populated,
as defaults are populated based on nil checks in Machine#ssh_info.

The default behavior of protobufs and mapstructure was making it so that
every field mentioned was getting its zero value returned. This was
preventing SSH from working properly in places like the `provision`
command.

This change is coupled with an update in the provider client within
vagrant-ruby which utilizes the new "hazzers" that get generated to
ensure the returned hash is sparsely populated.