Flesh out Basis-level fields
Thus far we've been getting by without using the Basis much on the Ruby
side, dangling most behavior off of the Project/Environment instead.
Here we add the necessary methods to allow Ruby to bootstrap a
Vagrant::Environment with a Basis client. This sets us up to fixup some
commands that need to execute successfully outside the context of a
Project.