Support for Component Options
In Legacy Vagrant, some types of plugins are registered with extra
metadata that is returned when the plugin is looked up.

This is different for each plugin type, and it's implemented by just
having the plugin manager return a tuple of `[plugin_class, other_args..]`
instead of just `plugin_class` directly.

Thie interface for these extra args is implicit: there's no type
checking or validation, just an assumption that plugin callers make
about the format of the metadata.

In Gogo, we can do better! Here we establish a `ComponentOptions()`
pattern where we are able to define a per-component struct that
establishes a strongly-typed set of metadata that can be used on the Go
side in a new `sdk.WithComponent(component, options)` call.

We implement `ProviderOptions` as the first guinea pig.

On the vagrant-ruby side we'll use ProviderOptions to honor priority,
box_optional, and defaultable when selecting a default provider.