Fix infinite loop when FindPath does not succeed
FindPath has a loop which is meant to terminate when it reaches the root
of a filesystem, however the condition it was using was assuming that
`==` would work between two paths. This never matches, so the loop would
keep chugging along forever if a file was not found.

To fix this, we extend the config.Path interface a bit and use the new
functions in the loop.

We also need to change the "not found" condition to return no error, as
both the function comment and the caller expect this to be the behavior.

Along the way, we update the existence check to match the recommendation
at https://stackoverflow.com/a/66405130