Add DefaultProvider() to Basis
In legacy, the single Environment#default_provider method handles both
inside-a-project and outside-a-project contexts, while in Gogo we have
more separation between Basis & Project. This means we need two
different implementations of DefaultProvider(), with the Basis-level one
being simpler because it doesn't have any config to consider.

The only time the Basis-level default_provider is called is from
Environment#guess_provider, so it doesn't need to take any options.