Wiring for Command plugin options
Command plugins are special as their metadata is passed along early in
vagrant-go's boot for the purposes of help. This uses a CommandInfo
struct that carries along metadata about the command.

In addition to being modeled as its own component.CommandOptions type,
the Primary field needs to be added to CommandInfo so it can be ferried
along with other help information, allowing us to honor the commands
that choose to be hidden by setting Primary to false.